Graf 91 add photo to graffiti and other (#172)
* GRAF-91 minor changes

* GRAF-91 removed MetadataService.ts

* GRAF-91 removed leftover imports

* GRAF-91 major prisma schema + sql migration overhaul i.e. v2_schema

* GRAF-91 working signup, pagaliau...

* GRAF-91 metadata service struct update

* GRAF-91 draft picture score implementation

* GRAF-91 pictureScore alg implementation

* GRAF-91 test fix

* GRAF-91 fix unhandled error

* GRAF-91 minor metadata & graffiti photo update

+++ src/user/user.controller.ts
@@ -7,6 +7,8 @@ import {
 	Param,
 	Delete,
 	Query,
+	UseGuards,
+	UnauthorizedException,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/request/create-user.dto';
@@ -14,6 +16,9 @@ import { UpdateUserDto } from './dto/request/update-user.dto';
 import UserMapper from './mapper/UserMapper';
 import { LikesEntry } from './dto/request/likesEntry.dto';
 import { ApiOperation } from '@nestjs/swagger';
+import { JwtAuthGuard } from '../auth/guards/jwt.guard';
+import GetCurrentUserId from '../auth/decorators/get-current-user-id.decorator';
+import Public from '../auth/decorators/public.decorator';
 
 @Controller('api/v1/user')
 export class UserController {
@@ -30,33 +35,36 @@ export class UserController {
 	// 	let entities = await this.userService.findAll();
 	// 	return UserMapper.toResponses(entities);
 	// }
+
 	@Get(':id')
 	async findById(@Param('id') id: string) {
-		console.log(id);
-		// if (isNaN(+id)) {
-		// 	return null;
-		// }
 		let entity = await this.userService.findById(+id);
-
-		if (entity !== null) {
-			return UserMapper.toResponse(entity);
-		}
-		return null;
+		return UserMapper.toResponse(entity);
 	}
 
+	@Public()
 	@Get('exist/:usernameOrEmail')
 	async existsAlready(@Param('usernameOrEmail') usernameOrEmail?: string) {
-		console.log('username', usernameOrEmail);
-
 		return await this.userService.usernameOrEmailExists(usernameOrEmail);
 	}
 
 	@Put(':id')
-	async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+	@UseGuards(JwtAuthGuard)
+	async update(
+		@Param('id') id: string,
+		@Body() updateUserDto: UpdateUserDto,
+		@GetCurrentUserId() userId: number,
+	) {
+		if (+id !== userId) {
+			throw new UnauthorizedException(
+				'You are not authorized to update this user',
+			);
+		}
 		let entity = await this.userService.update(+id, updateUserDto);
 		return UserMapper.toResponse(entity);
 	}
 
+	// remove this later?
 	@Put('/:id/likes/add')
 	@ApiOperation({ summary: 'Update user likes by id' })
 	async Likes(@Param('id') id: string, @Body() request: LikesEntry) {
@@ -64,6 +72,7 @@ export class UserController {
 		return UserMapper.toResponse(entity);
 	}
 
+	// remove this later?
 	@Put('/:id/likes/remove')
 	@ApiOperation({ summary: 'Update user likes by id' })
 	async removeLikesFromUser(
@@ -75,7 +84,13 @@ export class UserController {
 	}
 
 	@Delete(':id')
-	async delete(@Param('id') id: string) {
+	@UseGuards(JwtAuthGuard)
+	async delete(@Param('id') id: string, @GetCurrentUserId() userId: number) {
+		if (+id !== userId) {
+			throw new UnauthorizedException(
+				'You are not authorized to delete this user',
+			);
+		}
 		let entity = await this.userService.delete(+id);
 		return UserMapper.toResponse(entity);
 	}
